# Reduce cache contention on connection struct and branch mispredictions in connSocketEventHandler()