--- a/src/socket.c
+++ b/src/socket.c
@@ -249,8 +249,8 @@ static void connSocketEventHandler(struct aeEventLoop *el, int fd, void *clientD
     UNUSED(fd);
     connection *conn = clientData;
 
-    if (conn->state == CONN_STATE_CONNECTING &&
-            (mask & AE_WRITABLE) && conn->conn_handler) {
+    if (unlikely(conn->state == CONN_STATE_CONNECTING &&
+            (mask & AE_WRITABLE) && conn->conn_handler)) {
 
         int conn_error = anetGetError(conn->fd);
         if (conn_error) {