### Summary

This patch improves the layout of the struct connection to reduce cache contention and branch mispredictions under high concurrency, given connSocketEventHandler (and the connection struct) appeared as the top "hotspot" concerning L3 Bound/Contested Accesses, which occur when data written by one thread is read by another thread on a different core (see bellow):

![image](https://github.com/user-attachments/assets/f64ac19b-6246-4b65-a6e5-bcb1764dbacf)


This patch introduces:
- A dedicated cacheline for frequently updated fields (`state`, `flags`, `refs`, `iovcnt`).
- Explicit padding computed using `sizeof()` to make it portable across platforms.
- `__attribute__((aligned(CACHE_LINE_SIZE)))` to ensure alignment of the entire struct.
- `unlikely()` hint on the rare `CONN_STATE_CONNECTING` branch to reduce misprediction.

### VTune Highlights (Before → After)
| Metric | Before | After | Improvement |
|--------|--------|-------|-------------|
| CPI | 1.580 | 0.996 | -0.584 |
| L1 Bound | 65.4% | 20.5% | -44.9% |
| Contested Accesses | 45.4% | 8.5% | -36.9% |
| Bad Speculation | 5.3% | 2.1% | -3.2% |


### Client side Highlights (Before → After)
| Metric | Before | After | Improvement |
|--------|--------|-------|-------------|
| 100% GET use-case ops/sec | 898,471.80 | 911,186.22 | +1.4% |

### Confirmation that number of cachelines does not change 

Original struct: size: 72, cachelines: 2
```
# ( pahole  --structs ./src/redis-server | grep -B 20 -A 20 "struct connection" )

struct connection {
	ConnectionType *           type;                 /*     0     8 */
	ConnectionState            state;                /*     8     4 */
	int                        last_errno;           /*    12     4 */
	int                        fd;                   /*    16     4 */
	short int                  flags;                /*    20     2 */
	short int                  refs;                 /*    22     2 */
	short unsigned int         iovcnt;               /*    24     2 */

	/* XXX 6 bytes hole, try to pack */

	void *                     private_data;         /*    32     8 */
	struct aeEventLoop *       el;                   /*    40     8 */
	ConnectionCallbackFunc     conn_handler;         /*    48     8 */
	ConnectionCallbackFunc     write_handler;        /*    56     8 */
	/* --- cacheline 1 boundary (64 bytes) --- */
	ConnectionCallbackFunc     read_handler;         /*    64     8 */

	/* size: 72, cachelines: 2, members: 12 */
	/* sum members: 66, holes: 1, sum holes: 6 */
	/* last cacheline: 8 bytes */
};
```

Revised order and padding struct: size: 128, cachelines: 2

```
struct connection {
	ConnectionState            state;                /*     0     4 */
	short int                  flags;                /*     4     2 */
	short int                  refs;                 /*     6     2 */
	short unsigned int         iovcnt;               /*     8     2 */
	char                       _pad0[54];            /*    10    54 */
	/* --- cacheline 1 boundary (64 bytes) --- */
	ConnectionType *           type;                 /*    64     8 */
	int                        last_errno;           /*    72     4 */
	int                        fd;                   /*    76     4 */
	void *                     private_data;         /*    80     8 */
	struct aeEventLoop *       el;                   /*    88     8 */
	ConnectionCallbackFunc     conn_handler;         /*    96     8 */
	ConnectionCallbackFunc     write_handler;        /*   104     8 */
	ConnectionCallbackFunc     read_handler;         /*   112     8 */

	/* size: 128, cachelines: 2, members: 13 */
	/* padding: 8 */
} __attribute__((__aligned__(64)));
```

### Outcome
This change significantly reduces stalls from both **cache contention** and **branch mispredictions**, leading to improved performance under multi-threaded workloads without altering application logic.