@fcostaoliveira 2025-04-01
### CE Performance Automation : step 1 of 2 (build) STARTING...

This comment was automatically generated given a benchmark was triggered.
Started building at 2025-04-01 23:20:21.788988
You can check each build/benchmark progress in grafana:
   - git hash: 6fd8c07b363b6c0d8554024e11a6d3200db92628
   - git branch: filipecosta90:connection.hotpath
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*